refactored hocs into containers
